implemented 32-bit counter implementation

+++ include/afl-fuzz.h
@@ -1086,7 +1086,7 @@ void discover_word(u8 *ret, u64 *current, u64 *virgin);
 void discover_word(u8 *ret, u32 *current, u32 *virgin);
 #endif
 void init_count_class16(void);
-void minimize_bits(afl_state_t *, u8 *, u8 *);
+void minimize_bits(afl_state_t *, u8 *, u32 *);
 #ifndef SIMPLE_FILES
 u8 *describe_op(afl_state_t *, u8, size_t);
 #endif
